requirements.txt
Django==1.10
djangorestframework==3.7.3
PyJWT==1.5.3
pytz==2017.3
